improved the navbar
